+++ demos/petclinic/src/main/java/is/codion/framework/demos/petclinic/ui/VetSpecialtyEditPanel.java
@@ -38,7 +38,7 @@ public final class VetSpecialtyEditPanel extends EntityEditPanel {
 
 		createForeignKeyComboBox(VetSpecialty.VET_FK);
 		createForeignKeyComboBoxPanel(VetSpecialty.SPECIALTY_FK, this::createSpecialtyEditPanel)
-						.addButton(true);
+						.includeAddButton(true);
 
 		setLayout(gridLayout(2, 1));